Implemented user authentication so you have to login to add series and to edit or remove your own added series.

+++ week2/model.php
@@ -97,7 +97,7 @@ function get_breadcrumbs($breadcrumbs) {
  * @param array $navigation Array with as Key the page name and as Value the corresponding url
  * @return string html code that represents the navigation
  */
-function get_navigation($navigation){
+function get_navigation($navigation_tpl, $active_id){
     $navigation_exp = '
     <nav class="navbar navbar-expand-lg navbar-light bg-light">
     <a class="navbar-brand">Series Overview</a>
@@ -106,13 +106,14 @@ function get_navigation($navigation){
     </button>
     <div class="collapse navbar-collapse" id="navbarSupportedContent">
     <ul class="navbar-nav mr-auto">';
-    foreach ($navigation as $name => $info) {
-        if ($info[1]){
+    foreach ($navigation_tpl as $name => $info) {
+        $nav_info = na($name, $info);
+        if ($nav_info[0] == $active_id){
             $navigation_exp .= '<li class="nav-item active">';
-            $navigation_exp .= '<a class="nav-link" href="'.$info[0].'">'.$name.'</a>';
+            $navigation_exp .= '<a class="nav-link" href="'.$nav_info[1]['url'].'">'.$nav_info[1]['name'].'</a>';
         }else{
             $navigation_exp .= '<li class="nav-item">';
-            $navigation_exp .= '<a class="nav-link" href="'.$info[0].'">'.$name.'</a>';
+            $navigation_exp .= '<a class="nav-link" href="'.$nav_info[1]['url'].'">'.$nav_info[1]['name'].'</a>';
         }
 
         $navigation_exp .= '</li>';
@@ -126,16 +127,17 @@ function get_navigation($navigation){
 
 /**
  * Creats a Bootstrap table with a list of series
- * @param object $db pdo object
+ * @param PDO $pdo database object
  * @param array $series with series from the db
  * @return string
  */
-function get_serie_table($series){
+function get_serie_table($series, $pdo){
     $table_exp = '
     <table class="table table-hover">
     <thead
     <tr>
         <th scope="col">Series</th>
+        <th scope="col">Added by</th>
         <th scope="col"></th>
     </tr>
     </thead>
@@ -144,7 +146,8 @@ function get_serie_table($series){
         $table_exp .= '
         <tr>
             <th scope="row">'.$value['name'].'</th>
-            <td><a href="/DDWT18/week2/serie/?serie_id='.$value['id'].'" role="button" class="btn btn-primary">More info</a></td>
+            <td>'.(get_username($pdo, $value['user'])['firstname']." ".get_username($pdo, $value['user'])['lastname']).'</td>
+            <td><a href="/DDWT18/week2/serie/?serie_id='.$value['id'].'&user_id='.$value['user'].'" role="button" class="btn btn-primary">More info</a></td>
         </tr>
         ';
     }
@@ -211,6 +214,7 @@ function get_serieinfo($pdo, $serie_id){
  * @return string
  */
 function get_error($feedback){
+    $feedback = json_decode($feedback, True);
     $error_exp = '
         <div class="alert alert-'.$feedback['type'].'" role="alert">
             '.$feedback['message'].'
@@ -224,7 +228,7 @@ function get_error($feedback){
  * @param array $serie_info post array
  * @return array with message feedback
  */
-function add_serie($pdo, $serie_info){
+function add_serie($pdo, $serie_info, $userid){
     /* Check if all fields are set */
     if (
         empty($serie_info['Name']) or
@@ -258,12 +262,13 @@ function add_serie($pdo, $serie_info){
     }
 
     /* Add Serie */
-    $stmt = $pdo->prepare("INSERT INTO series (name, creator, seasons, abstract) VALUES (?, ?, ?, ?)");
+    $stmt = $pdo->prepare("INSERT INTO series (name, creator, seasons, abstract, user) VALUES (?, ?, ?, ?, ?)");
     $stmt->execute([
         $serie_info['Name'],
         $serie_info['Creator'],
         $serie_info['Seasons'],
-        $serie_info['Abstract']
+        $serie_info['Abstract'],
+        $userid
     ]);
     $inserted = $stmt->rowCount();
     if ($inserted ==  1) {
@@ -286,7 +291,13 @@ function add_serie($pdo, $serie_info){
  * @param array $serie_info post array
  * @return array
  */
-function update_serie($pdo, $serie_info){
+function update_serie($pdo, $serie_info, $userid, $user){
+    if ($userid != $user) {
+        return [
+            'type' => 'warning',
+            'message' => 'You are not allowed to edit this series'
+        ];
+    }
     /* Check if all fields are set */
     if (
         empty($serie_info['Name']) or
@@ -326,28 +337,29 @@ function update_serie($pdo, $serie_info){
         ];
     }
 
-    /* Update Serie */
-    $stmt = $pdo->prepare("UPDATE series SET name = ?, creator = ?, seasons = ?, abstract = ? WHERE id = ?");
-    $stmt->execute([
-        $serie_info['Name'],
-        $serie_info['Creator'],
-        $serie_info['Seasons'],
-        $serie_info['Abstract'],
-        $serie_info['serie_id']
-    ]);
-    $updated = $stmt->rowCount();
-    if ($updated ==  1) {
-        return [
-            'type' => 'success',
-            'message' => sprintf("Series '%s' was edited!", $serie_info['Name'])
-        ];
-    }
-    else {
-        return [
-            'type' => 'warning',
-            'message' => 'The series was not edited. No changes were detected'
-        ];
-    }
+
+        /* Update Serie */
+        $stmt = $pdo->prepare("UPDATE series SET name = ?, creator = ?, seasons = ?, abstract = ?, user = ? WHERE id = ?");
+        $stmt->execute([
+            $serie_info['Name'],
+            $serie_info['Creator'],
+            $serie_info['Seasons'],
+            $serie_info['Abstract'],
+            $user,
+            $serie_info['serie_id']
+        ]);
+        $updated = $stmt->rowCount();
+        if ($updated == 1) {
+            return [
+                'type' => 'success',
+                'message' => sprintf("Series '%s' was edited!", $serie_info['Name'])
+            ];
+        } else {
+            return [
+                'type' => 'warning',
+                'message' => 'The series was not edited. No changes were detected'
+            ];
+        }
 }
 
 /**
@@ -359,23 +371,30 @@ function update_serie($pdo, $serie_info){
 function remove_serie($pdo, $serie_id){
     /* Get series info */
     $serie_info = get_serieinfo($pdo, $serie_id);
+    $current_user = get_user_id();
+    if ($current_user != $serie_info['user']) {
+        return [
+            'type' => 'danger',
+            'message' => 'You are not allowed to delete this series'
+        ];
+    }
 
-    /* Delete Serie */
-    $stmt = $pdo->prepare("DELETE FROM series WHERE id = ?");
-    $stmt->execute([$serie_id]);
-    $deleted = $stmt->rowCount();
-    if ($deleted ==  1) {
-        return [
-            'type' => 'success',
-            'message' => sprintf("Series '%s' was removed!", $serie_info['name'])
-        ];
-    }
-    else {
-        return [
-            'type' => 'warning',
-            'message' => 'An error occurred. The series was not removed.'
-        ];
-    }
+        /* Delete Serie */
+        $stmt = $pdo->prepare("DELETE FROM series WHERE id = ?");
+        $stmt->execute([$serie_id]);
+        $deleted = $stmt->rowCount();
+        if ($deleted == 1) {
+            return [
+                'type' => 'success',
+                'message' => sprintf("Series '%s' was removed!", $serie_info['name'])
+            ];
+        } else {
+            return [
+                'type' => 'warning',
+                'message' => 'An error occurred. The series was not removed.'
+            ];
+        }
+
 }
 
 /**
@@ -408,6 +427,165 @@ function get_user_id(){
     session_start();
     if (isset($_SESSION['user_id'])){
         return $_SESSION['user_id'];
+    } else {
+        return False;
+    }
+}
+
+/**
+ * Get the full name of a user corresponding with a specific id
+ * @param PDO $pdo database object
+ * @return mixed
+ */
+function get_username($pdo, $user_id) {
+    $stmt = $pdo->prepare("SELECT firstname, lastname FROM users WHERE id=?");
+    $stmt->execute([$user_id]);
+    $username = $stmt->fetch();
+    return $username;
+}
+
+/**
+ * Count the current number of users in the database
+ * @param PDO $pdo database object
+ * @return mixed
+ */
+function count_users($pdo) {
+    $stmt = $pdo->prepare("SELECT * FROM users");
+    $stmt->execute();
+    $users = $stmt->rowCount();
+    return $users;
+}
+
+/**
+ * Count the current number of users in the database
+ * @param PDO $pdo database object
+ * @return mixed
+ */
+function register_user($pdo, $form_data) {
+    if (
+        empty($form_data['username']) or
+        empty($form_data['password']) or
+        empty($form_data['firstname']) or
+        empty($form_data['lastname'])
+    ) {
+        return [
+            'type' => 'danger',
+            'message' => 'Please enter an username, a password, your first name and your last name'
+        ];
+    }
+
+    /* Check if user already exists */
+    try {
+        $stmt = $pdo->prepare('SELECT * FROM users WHERE username = ?');
+        $stmt->execute([$form_data['username']]);
+        $user_exists = $stmt->rowCount();
+    } catch (\PDOException $e) {
+        return [
+            'type' => 'danger',
+            'message' => sprintf('There was an error: %s', $e->getMessage())
+        ];
+    }
+
+    /* Return error message for existing username */
+    if ( !empty($user_exists) ) {
+        return [
+            'type' => 'danger',
+            'message' => 'The username you entered does already exists!'
+        ];
+    }
+    /* Hash password */
+    $password = password_hash($form_data['password'], PASSWORD_DEFAULT);
+    /* Save user to the database */
+    try {
+        $stmt = $pdo->prepare('INSERT INTO users (username, password, firstname,
+lastname) VALUES (?, ?, ?, ?)');
+        $stmt->execute([$form_data['username'], $password, $form_data['firstname'],
+            $form_data['lastname']]);
+        $user_id = $pdo->lastInsertId();
+    } catch (PDOException $e) {
+        return [
+            'type' => 'danger',
+            'message' => sprintf('There was an error: %s', $e->getMessage())
+        ];
+    }
+
+    /* Login user and redirect */
+    session_start();
+    $_SESSION['user_id'] = $user_id;
+    $feedback = [
+        'type' => 'success',
+        'message' => sprintf('%s, your account was successfully
+created!', (get_username($pdo, $_SESSION['user_id'])['firstname']) . " " . get_username($pdo, $_SESSION['user_id'])['lastname'])
+    ];
+    redirect(sprintf('/DDWT18/week2/myaccount/?error_msg=%s',
+        json_encode($feedback)));
+}
+
+/**
+ * Count the current number of users in the database
+ * @param PDO $pdo database object
+ * @return mixed
+ */
+function login_user($pdo, $form_data)
+{
+    if (empty($form_data['username']) or
+        empty($form_data['password'])) {
+        return [
+            'type' => 'warning',
+            'message' => 'Please fill in both your username and password'
+        ];
+    }
+    /* Check if the username exists*/
+    try {
+        $stmt = $pdo->prepare('SELECT * FROM users WHERE username=?');
+        $stmt->execute([$form_data['username']]);
+        $user_info = $stmt->fetch();
+    } catch (PDOException $e) {
+        return [
+            'type' => 'danger',
+            'message' => sprintf('There was an error: %s', $e->getMessage())
+        ];
+    }
+    /*Return an error is the username does not exist*/
+    if (empty($user_info)) {
+        return [
+            'type' => 'warning',
+            'message' => 'The username you entered does not exist.'
+        ];
+    }
+    /* Check password */
+    if (!password_verify($form_data['password'], $user_info['password'])) {
+        return [
+            'type' => 'danger',
+            'message' => 'The password you entered is incorrect!'
+        ];
+    } else {
+        session_start();
+        $_SESSION['user_id'] = $user_info['id'];
+        $feedback = [
+            'type' => 'success',
+            'message' => sprintf('%s, you were logged in successfully!',
+                (get_username($pdo, $_SESSION['user_id'])['firstname']) . " " . get_username($pdo, $_SESSION['user_id'])['lastname'])
+        ];
+        redirect(sprintf('/DDWT18/week2/myaccount/?error_msg=%s&user_id=%s',
+            json_encode($feedback), $_SESSION['user_id']));
+    }
+}
+
+function logout_user($pdo) {
+    session_start();
+    session_destroy();
+    $feedback = [
+        'type' => 'success',
+        'message' => sprintf('%s, you have been succesfully logged out!', (get_username($pdo, $_SESSION['user_id'])['firstname'])." ".get_username($pdo, $_SESSION['user_id'])['lastname'])
+    ];
+    return $feedback;
+}
+
+function check_login(){
+    session_start();
+    if (isset($_SESSION['user_id'])){
+        return True;
     } else {
         return False;
     }
